SummerProject/requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.4
Django==2.2.3
django-handlers==0.1.1
djangorestframework==3.10.2
idna==2.8
importlib-metadata==0.19
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.1
pkg-resources==0.0.0
pluggy==0.12.0
psycopg2==2.8.3
py==1.8.0
pylint==2.3.1
pyparsing==2.4.2
pytest==5.0.1
pytest-cov==2.7.1
pytest-django==3.5.1
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.5.2
